refactor(accounting): extract components, fix colors, safe redirect, add DB indexes
P2-1 — Extract sub-components from the three monolith clients
The reconciliation, exports and entries pages were each ~640 LOC of
table + form + mutation orchestration crammed into a single file. Each
one is now a thin parent (hooks + state + mutations) wrapping a set of
dumb sub-components under ./components/.

ReconciliationClient.tsx  (636 → 182 LOC)
  components/ReconciliationStats.tsx     — 3 KPI cards + progress bar
  components/ReconciliationFilters.tsx   — connection tabs + status tabs
  components/EntryMatchCard.tsx          — suggested-entry chip
  components/TransactionCard.tsx         — full bank-tx card (owns its
                                            own local drawer state but
                                            every mutation is a prop)

ExportsPageClient.tsx     (647 → 498 LOC)
  components/CahierExportPanel.tsx       — Récapitulatif annuel + 3 KPI
  components/GrandLivreExportPanel.tsx   — wraps ExportCard PDF+CSV
  components/BalanceExportPanel.tsx      — wraps ExportCard PDF+CSV
  components/FECExportPanel.tsx          — FEC preview + download block

EntriesPageClient.tsx     (622 → 354 LOC)
  components/EntryFilters.tsx            — search + journal + status +
                                            source + date range selects
  components/BulkActions.tsx             — "validate selected" chip
  components/EntryRow.tsx                — <EntryRow> desktop <tr> +
                                            <EntryCard> mobile card,
                                            sharing the draft/debit/
                                            credit helpers (re-exported
                                            for the parent's totals)

Rules followed:
- Every sub-component has a typed props interface
- Hooks stay in the parent (useAccountingEntries, useReconciliation,
  useBankConnections, useMutation, toasts)
- Mutations are passed down as callbacks (onMatch, onIgnore,
  onCategorize, onValidateSelected, onDownload, onLoadPreview)
- Zero visible rendering change: same markup, same order, same styling

P2-2 — Remove hardcoded hex colors
Swap every #2563EB / #1B2A6B in the accounting module for design-system
tokens. The gradient treasury card and every "primary hover / focus"
ring now go through hsl(var(--primary)) and its /90 /50 opacity steps:
- BankAccountsClient.tsx:  from-[#2563EB] to-[#1B2A6B] → from-primary
  to-primary/70, hover:bg-[#1B2A6B] → hover:bg-primary/90, text-white
  → text-primary-foreground.
- ConnectBankClient.tsx and ReconciliationClient.tsx: focus:ring-[#2563EB]
  → focus:ring-primary/50, hover:bg-[#1B2A6B] → hover:bg-primary/90.
grep "#[0-9A-Fa-f]\{6\}" on app/owner/accounting/**/*.tsx now returns 0
results.

P2-3 — Validate bank OAuth redirect target
ConnectBankClient was doing `window.location.href = authLink` on
whatever URL the backend returned. Added an inline isTrustedBankAuthLink()
helper + ALLOWED_BANK_AUTH_HOSTS whitelist
('ob.nordigen.com', 'bankaccountdata.gocardless.com') that rejects
non-https and any other host, with a destructive toast explaining the
situation. The submit button then simply bails out instead of
redirecting. Prevents open-redirect / phishing if the /bank/connections
response is ever tampered with.

P2-4 — Missing indexes migration
New migration supabase/migrations/20260410100000_accounting_missing_indexes.sql
adds the composite / trigram indexes the audit flagged. The task's
suggested columns don't all exist on the real schema, so the indexes
target the actual hot query paths:
  - idx_entries_exercise_date       : (exercise_id, entry_date DESC)
    for grand-livre traversal by exercise.
  - idx_entry_lines_account_entry   : (account_number, entry_id) on
    accounting_entry_lines (no exercise_id column lives here, it's on
    the parent accounting_entries; composite with entry_id is what
    actually speeds up the join).
  - idx_bank_tx_connection_status_date : (connection_id,
    reconciliation_status, transaction_date DESC) — bank_transactions
    has no entity_id (entity lives on bank_connections), so the
    composite keys off connection_id, which is what
    /api/accounting/bank/reconciliation filters on.
  - idx_entries_label_trgm + idx_entries_piece_ref_trgm : GIN trigram
    indexes over accounting_entries.label and piece_ref to make the
    EntriesPageClient ILIKE search selective.
pg_trgm is already enabled globally by
20240101000000_initial_schema.sql so no CREATE EXTENSION is needed
here. Every statement is CREATE INDEX IF NOT EXISTS (idempotent).

tsc --noEmit: 508 errors (unchanged baseline, no regression).
Monolith LOC: 636+647+622 (1905) → 182+498+354 (1034) — 46% smaller
parents plus 11 reusable sub-components.
